+++ environment.yml
@@ -1,4 +1,4 @@
-name: core
+name: base
 channels:
   - conda-forge
   - defaults
