+++ src/caf/base/__init__.py
@@ -1,0 +1,9 @@
+"""Package description."""
+
+from caf.base import data_structures, segmentation, segments, zoning
+from caf.base.data_structures import DVector
+from caf.base.segmentation import Segmentation, SegmentationInput
+from caf.base.segments import Segment
+from caf.base.zoning import BalancingZones, ZoningSystem
+
+from ._version import __version__
