+++ src/caf/base/seg_translations/sic_2_digit_to_sic_1_digit.csv
