+++ src/caf/base/segments.py
