+++ src/caf/base/segments/aws.yml
