+++ src/caf/base/segments/car_availability.yml
