+++ src/caf/base/segments/economic_status.yml
