+++ src/caf/base/segments/hh_type.yml
