+++ src/caf/base/segments/pop_econ.yml
