+++ src/caf/base/segments/soc.yml
