--- src/caf/core/__init__.py
@@ -1,9 +1,0 @@
-"""Package description."""
-
-from caf.core import data_structures, segmentation, segments, zoning
-from caf.core.data_structures import DVector
-from caf.core.segmentation import Segmentation, SegmentationInput
-from caf.core.segments import Segment
-from caf.core.zoning import BalancingZones, ZoningSystem
-
-from ._version import __version__
